+++ test/report_test.exs
@@ -1,5 +1,5 @@
 defmodule Etop.ReportTest do
-  use ExUnit.Case, async: true
+  use ExUnit.Case
 
   import ExUnit.CaptureIO